requirements.txt
pycparser
pynacl
toml
colorama
watchdog
levenshtein
cxxfilt
GitPython
colour
splat64[mips]>=0.34.0,<1.0.0